introduced a layer of indirection between timestamp search logic and ethereum
